Adding Entities

+++ src/main/java/com/adsima/spring/dto/CompanyReadDto.java
@@ -1,5 +1,5 @@
 package com.adsima.spring.dto;
 
-public record CompanyReadDto(Long id)
+public record CompanyReadDto(Integer id)
 {
 }
